requirements.txt
requests
beautifulsoup4
mysql-connector-python
pymilvus
sentence-transformers
transformers
spacy
scikit-learn

-e .
